Get Kube-arango operator working (#51)
This commit gets the kube-arangodb operator going and does it by
splitting the config into folders.
The idea here is that by moving all the manifests for things in the
api/istio-system/frontend namespace into their own folders with their
own kustomization.yaml, we can then apply a namespace across all the
objects those files produce. This is a big deal with Istio (given the
number of objects it creates), and even more so with the Kube-arangodb
operator, since it actually specifies the default namespace.